# internal/models/user_model.go
package models

import (
	"time"

	"go.mongodb.org/mongo-driver/bson/primitive"
)

type UserRole string

const (
	RoleUser  UserRole = "USER"
	RoleAdmin UserRole = "ADMIN"
)

type User struct {
	ID         primitive.ObjectID `bson:"_id,omitempty" json:"id"`
	GoogleID   string             `bson:"google_id" json:"google_id"`
	ImgProfile string             `bson:"img_profile" json:"img_profile"`
	Email      string             `bson:"email" json:"email"`
	Name       string             `bson:"name" json:"name"`
	Role       UserRole           `bson:"role" json:"role"`
	CreatedAt  time.Time          `bson:"created_at" json:"created_at"`
	UpdatedAt  time.Time          `bson:"updated_at" json:"updated_at"`
}